Changed various properties and parameters to be Int32 rather than Int64 as was unnecessary.
git-svn-id: https://flickrnet.svn.codeplex.com/svn@58915 81e6e737-5a20-4d0f-a97c-cb01297b46e0

+++ FlickrNet/Categories.cs
@@ -74,7 +74,7 @@ namespace FlickrNet
 		/// The number of groups found within the category.
 		/// </summary>
 		[XmlAttribute("count", Form=XmlSchemaForm.Unqualified)]
-		public long GroupCount;
+		public int GroupCount;
 	}
 
 }
